build: un-share single-caller dependency-update functions
update_dependencies had exactly one call site (in update_version.py's
loop) so it's inlined there directly. update_patch_dependencies has
two call sites, both in update_patch_version.py, so it stays a
function but moves out of version_files.py into the file that
actually uses it. version_files.py keeps only what's genuinely shared
between both entry points.
